UI and Clean up
