+++ .vscode/extensions.json
@@ -1,3 +1,4 @@
 {
-  "recommendations": ["tauri-apps.tauri-vscode", "rust-lang.rust-analyzer"]
+  "recommendations": ["tauri-apps.tauri-vscode", "rust-lang.rust-analyzer"],
+  "css.lint.unknownAtRules": "ignore",
 }
