+++ src/App.css
@@ -5,21 +5,27 @@
   font-family: Inter, Avenir, Helvetica, Arial, sans-serif;
 }
 
-button{
+@theme {
+  --color-main-dark-bg: var(bg-gray-900);
+}
+
+button {
   cursor: pointer;
 }
 /* Small helper to ensure full-height pages if needed */
-html, body, #root {
+html,
+body,
+#root {
   height: 100%;
 }
 
-::-webkit-scrollbar{
+::-webkit-scrollbar {
   height: 10px;
   width: 4px;
 }
 
-::-webkit-scrollbar-thumb{
-  background-color: rgb(135, 135, 136);
+::-webkit-scrollbar-thumb {
+  background-color: royalblue;
   width: 4px;
   border-radius: 999px;
 }
