+++ src/EmojiData/SymbolsAndASCII.ts
@@ -1,0 +1,526 @@
+import { Emojies } from "../types/app.types";
+
+export const symbolEmoticonArray: Emojies[] = [
+  //    currency
+  {
+    label: "USD",
+    type: "Currency",
+    emoji: "$",
+    keywords: ["dollar", "usd", "money", "cash", "finance"],
+  },
+  {
+    label: "Euro",
+    type: "Currency",
+    emoji: "€",
+    keywords: ["euro", "eur", "europe", "money", "cash"],
+  },
+  {
+    label: "British Pound",
+    type: "Currency",
+    emoji: "£",
+    keywords: ["pound", "gbp", "sterling", "money", "uk"],
+  },
+  {
+    label: "Yen / Yuan",
+    type: "Currency",
+    emoji: "¥",
+    keywords: ["yen", "yuan", "jpy", "cny", "japan", "china"],
+  },
+  {
+    label: "INR",
+    type: "Currency",
+    emoji: "₹",
+    keywords: ["rupee", "inr", "india", "money"],
+  },
+  {
+    label: "Russian Ruble",
+    type: "Currency",
+    emoji: "₽",
+    keywords: ["ruble", "rub", "russia", "money"],
+  },
+  {
+    label: "South Korean Won",
+    type: "Currency",
+    emoji: "₩",
+    keywords: ["won", "krw", "korea", "money"],
+  },
+  {
+    label: "Thai Baht",
+    type: "Currency",
+    emoji: "฿",
+    keywords: ["baht", "thb", "thailand", "money"],
+  },
+  {
+    label: "Turkish Lira",
+    type: "Currency",
+    emoji: "₺",
+    keywords: ["lira", "try", "turkey", "money"],
+  },
+  {
+    label: "Swiss Franc",
+    type: "Currency",
+    emoji: "₣",
+    keywords: ["franc", "chf", "switzerland", "money", "cash"],
+  },
+  {
+    label: "Israeli Shekel",
+    type: "Currency",
+    emoji: "₪",
+    keywords: ["shekel", "ils", "israel", "money"],
+  },
+  {
+    label: "VN Dong",
+    type: "Currency",
+    emoji: "₫",
+    keywords: ["dong", "vnd", "vietnam", "money"],
+  },
+  {
+    label: "Hryvnia",
+    type: "Currency",
+    emoji: "₴",
+    keywords: ["hryvnia", "uah", "ukraine", "money"],
+  },
+  {
+    label: "Peso",
+    type: "Currency",
+    emoji: "₱",
+    keywords: ["peso", "php", "philippines", "money"],
+  },
+  {
+    label: "Cedi",
+    type: "Currency",
+    emoji: "₵",
+    keywords: ["cedi", "ghs", "ghana", "money"],
+  },
+  {
+    label: "Tugrik",
+    type: "Currency",
+    emoji: "₮",
+    keywords: ["tugrik", "mnt", "mongolia", "money"],
+  },
+  {
+    label: "Lao Kip",
+    type: "Currency",
+    emoji: "₭",
+    keywords: ["kip", "lak", "laos", "money"],
+  },
+  {
+    label: "Sucre",
+    type: "Currency",
+    emoji: "₳",
+    keywords: ["sucre", "ecuador", "historic", "money"],
+  },
+  // symbols
+  {
+    label: "copyright",
+    type: "Symbols",
+    emoji: "©",
+    keywords: ["copyright", "legal", "owner", "license", "©"],
+  },
+  {
+    label: "registered",
+    type: "Symbols",
+    emoji: "®",
+    keywords: ["registered", "trademark", "brand", "legal", "®"],
+  },
+  {
+    label: "trademark",
+    type: "Symbols",
+    emoji: "™",
+    keywords: ["trademark", "tm", "brand", "legal", "™"],
+  },
+  {
+    label: "check mark",
+    type: "Symbols",
+    emoji: "✓",
+    keywords: ["check", "done", "yes", "correct", "verify", "✓"],
+  },
+  {
+    label: "bullet point",
+    type: "Symbols",
+    emoji: "•",
+    keywords: ["bullet", "point", "dot", "list", "•"],
+  },
+  {
+    label: "degree",
+    type: "Symbols",
+    emoji: "°",
+    keywords: ["degree", "temperature", "weather", "angle", "°"],
+  },
+  {
+    label: "infinity",
+    type: "Symbols",
+    emoji: "∞",
+    keywords: ["infinity", "forever", "loop", "math", "∞"],
+  },
+  {
+    label: "section",
+    type: "Symbols",
+    emoji: "§",
+    keywords: ["section", "paragraph", "legal", "law", "§"],
+  },
+  {
+    label: "currency generic",
+    type: "Symbols",
+    emoji: "¤",
+    keywords: ["currency", "money", "symbol", "price", "¤"],
+  },
+  // --- Arithmetic / Math ---
+  {
+    label: "Plus",
+    type: "Arithmetic",
+    emoji: "+",
+    keywords: ["add", "plus", "sum", "increase", "positive"],
+  },
+  {
+    label: "Minus",
+    type: "Arithmetic",
+    emoji: "-",
+    keywords: ["subtract", "minus", "difference", "decrease", "negative"],
+  },
+  {
+    label: "Times",
+    type: "Arithmetic",
+    emoji: "×", // Unicode 'Multiplication Sign'
+    keywords: ["multiply", "times", "product", "cross"],
+  },
+  {
+    label: "Divide",
+    type: "Arithmetic",
+    emoji: "÷", // Unicode 'Division Sign'
+    keywords: ["divide", "quotient", "ratio"],
+  },
+  {
+    label: "Equals",
+    type: "Arithmetic",
+    emoji: "=",
+    keywords: ["equal", "result", "is", "equality"],
+  },
+  {
+    label: "Not Equal",
+    type: "Arithmetic",
+    emoji: "≠",
+    keywords: ["not equal", "unequal", "difference"],
+  },
+  {
+    label: "Approx",
+    type: "Arithmetic",
+    emoji: "≈",
+    keywords: ["approximate", "almost equal", "estimate"],
+  },
+  {
+    label: "Less Than",
+    type: "Arithmetic",
+    emoji: "<",
+    keywords: ["smaller", "less", "lower"],
+  },
+  {
+    label: "Greater Than",
+    type: "Arithmetic",
+    emoji: ">",
+    keywords: ["bigger", "greater", "higher"],
+  },
+  {
+    label: "Equal To",
+    type: "Arithmetic",
+    emoji: "≤",
+    keywords: ["less equal", "at most"],
+  },
+  {
+    label: "Greater Equal",
+    type: "Arithmetic",
+    emoji: "≥",
+    keywords: ["greater equal", "at least"],
+  },
+  {
+    label: "Percent",
+    type: "Arithmetic",
+    emoji: "%",
+    keywords: ["percentage", "rate", "ratio", "out of 100"],
+  },
+  {
+    label: "Permille",
+    type: "Arithmetic",
+    emoji: "‰",
+    keywords: ["per thousand", "basis points", "rate"],
+  },
+  {
+    label: "Infinity",
+    type: "Arithmetic",
+    emoji: "∞",
+    keywords: ["endless", "limitless", "math", "forever"],
+  },
+  {
+    label: "Square Root",
+    type: "Arithmetic",
+    emoji: "√",
+    keywords: ["root", "radical", "math"],
+  },
+  {
+    label: "Power",
+    type: "Arithmetic",
+    emoji: "^",
+    keywords: ["caret", "power", "exponent", "to the power of"],
+  },
+  {
+    label: "Pi",
+    type: "Arithmetic",
+    emoji: "π",
+    keywords: ["pi", "constant", "geometry", "circle", "3.14"],
+  },
+  {
+    label: "Sigma",
+    type: "Arithmetic",
+    emoji: "Σ",
+    keywords: ["sigma", "sum", "total", "series", "math"],
+  },
+  {
+    label: "Integral",
+    type: "Arithmetic",
+    emoji: "∫",
+    keywords: ["calculus", "integration", "area", "math"],
+  },
+  {
+    label: "Partial",
+    type: "Arithmetic",
+    emoji: "∂",
+    keywords: ["derivative", "calculus", "partial"],
+  },
+  {
+    label: "Delta",
+    type: "Arithmetic",
+    emoji: "Δ",
+    keywords: ["delta", "change", "difference", "triangle"],
+  },
+  {
+    label: "varies as",
+    type: "Arithmetic",
+    emoji: "∝",
+    keywords: ["proportional", "varies as", "relation"],
+  },
+  {
+    label: "proof",
+    type: "Arithmetic",
+    emoji: "∴",
+    keywords: ["therefore", "conclusion", "logic", "proof"],
+  },
+  {
+    label: "Since",
+    type: "Arithmetic",
+    emoji: "∵",
+    keywords: ["because", "since", "reason", "logic"],
+  },
+  {
+    label: "Diameter",
+    type: "Arithmetic",
+    emoji: "∅",
+    keywords: ["empty set", "null", "diameter", "math"],
+  },
+  {
+    label: "element of",
+    type: "Arithmetic",
+    emoji: "∈",
+    keywords: ["element of", "belongs to", "set theory"],
+  },
+  {
+    label: "Angle",
+    type: "Arithmetic",
+    emoji: "∠",
+    keywords: ["angle", "geometry", "corner"],
+  },
+  {
+    label: "right angle",
+    type: "Arithmetic",
+    emoji: "⊥",
+    keywords: ["perpendicular", "orthogonal", "right angle"],
+  },
+  {
+    label: "Fn",
+    type: "Arithmetic",
+    emoji: "ƒ",
+    keywords: ["function", "math", "mapping"],
+  },
+  // quaso
+  { label: "happy", type: "Quaso", emoji: "(＾▽＾)", keywords: ["happy", "smile", "joy"] },
+  { label: "joy and playful", type: "Quaso", emoji: "(≧◡≦)", keywords: ["joy", "playful", "cute"] },
+  { label: "love", type: "Quaso", emoji: "(♡˙︶˙♡)", keywords: ["love", "heart", "cute"] },
+  { label: "kiss", type: "Quaso", emoji: "( ˘ ³˘)♥", keywords: ["kiss", "love", "affection"] },
+  { label: "sad", type: "Quaso", emoji: "(；︵；)", keywords: ["sad", "cry", "upset"] },
+  { label: "crying", type: "Quaso", emoji: "(T_T)", keywords: ["crying", "tears", "sad"] },
+  { label: "depressed", type: "Quaso", emoji: "(︶︹︶)", keywords: ["depressed", "down", "sad"] },
+  { label: "angry", type: "Quaso", emoji: "(╬ಠ益ಠ)", keywords: ["angry", "mad", "rage"] },
+  { label: "annoyed", type: "Quaso", emoji: "(¬_¬)", keywords: ["annoyed", "unimpressed", "meh"] },
+  {
+    label: "confused",
+    type: "Quaso",
+    emoji: "(・_・ヾ",
+    keywords: ["confused", "thinking", "unsure"],
+  },
+  {
+    label: "thinking",
+    type: "Quaso",
+    emoji: "(￢_￢)",
+    keywords: ["thinking", "suspicious", "hmm"],
+  },
+  { label: "surprised", type: "Quaso", emoji: "(⊙_⊙)", keywords: ["surprised", "shocked", "wow"] },
+  { label: "shrug", type: "Quaso", emoji: "¯\\_(ツ)_/¯", keywords: ["shrug", "idk", "whatever"] },
+  {
+    label: "table flip",
+    type: "Quaso",
+    emoji: "(╯°□°）╯︵ ┻━┻",
+    keywords: ["table flip", "angry", "rage"],
+  },
+  { label: "dead", type: "Quaso", emoji: "(x_x)", keywords: ["dead", "knocked out", "fail"] },
+  { label: "cute bear", type: "Quaso", emoji: "ʕ•ᴥ•ʔ", keywords: ["bear", "cute", "animal"] },
+  { label: "cute cat", type: "Quaso", emoji: "(=^･ω･^=)", keywords: ["cat", "cute", "animal"] },
+  {
+    label: "victory",
+    type: "Quaso",
+    emoji: "(ง •̀_•́)ง",
+    keywords: ["fight", "victory", "determined"],
+  },
+  { label: "strong", type: "Quaso", emoji: "ᕙ(⇀‸↼‶)ᕗ", keywords: ["strong", "muscle", "power"] },
+  {
+    label: "Table Flip",
+    type: "Quaso",
+    emoji: "(╯°□°）╯︵ ┻━┻",
+    keywords: ["table flip", "rage", "frustration", "anger"],
+  },
+  {
+    label: "blushing",
+    type: "Quaso",
+    emoji: "(⁄ ⁄>⁄ ▽ ⁄<⁄ ⁄)",
+    keywords: ["blush", "shy", "embarrassed"],
+  },
+  { label: "winking", type: "Quaso", emoji: "(^.~)", keywords: ["wink", "flirt", "playful"] },
+  { label: "sleeping", type: "Quaso", emoji: "(－_－) zzZ", keywords: ["sleep", "tired", "rest"] },
+  { label: "dance", type: "Quaso", emoji: "└(＾＾)┐", keywords: ["dance", "music", "happy"] },
+  { label: "excited", type: "Quaso", emoji: "o(>ω<)o", keywords: ["excited", "happy", "yay"] },
+  { label: "scared", type: "Quaso", emoji: "(ノдヽ)", keywords: ["scared", "hide", "fear"] },
+  { label: "pouting", type: "Quaso", emoji: " ( ￣＾￣)", keywords: ["pout", "grumpy", "mad"] },
+  {
+    label: "star eyes",
+    type: "Quaso",
+    emoji: "(☆▽☆)",
+    keywords: ["stars", "amazing", "impressed"],
+  },
+  { label: "hug", type: "Quaso", emoji: "(づ◡﹏◡)づ", keywords: ["hug", "cuddle", "comfort"] },
+  {
+    label: "lenny face",
+    type: "Quaso",
+    emoji: "( ͡° ͜ʖ ͡°)",
+    keywords: ["lenny", "smirk", "suspicious"],
+  },
+  { label: "dog", type: "Quaso", emoji: "V●ᴥ●V", keywords: ["dog", "puppy", "animal"] },
+  {
+    label: "sparkle happy",
+    type: "Quaso",
+    emoji: " (✧∀✧)/",
+    keywords: ["sparkle", "happy", "fabulous"],
+  },
+  { label: "worried", type: "Quaso", emoji: "(ーー;)", keywords: ["worried", "nervous", "sweat"] },
+  { label: "no thanks", type: "Quaso", emoji: "(￣ω￣)ゞ", keywords: ["salute", "ok", "bye"] },
+  {
+    label: "cool",
+    type: "Quaso",
+    emoji: "(⌐■_■)",
+    keywords: ["cool", "sunglasses", "deal with it"],
+  },
+  {
+    label: "money eyes",
+    type: "Quaso",
+    emoji: " [̲̅$̲̅(̲̅ ͡° ͜ʖ ͡°̲̅)̲̅$̲̅]",
+    keywords: ["money", "rich", "cash"],
+  },
+  {
+    label: "speechless",
+    type: "Quaso",
+    emoji: "(O_O;)",
+    keywords: ["speechless", "wow", "awkward"],
+  },
+  { label: "flexing", type: "Quaso", emoji: "ᕦ(ò_ó)ᕤ", keywords: ["strong", "flex", "tough"] },
+  {
+    label: "magical",
+    type: "Quaso",
+    emoji: "╰( ͡° ͜ʖ ͡° )つ──☆*:・ﾟ",
+    keywords: ["magic", "wizard", "sparkle"],
+  },
+  { label: "running", type: "Quaso", emoji: "ε=ε=┌(;￣▽￣)┘", keywords: ["run", "fast", "hurry"] },
+  // --- Happy/Positive ---
+  {
+    label: "Basic Smile",
+    type: "Sy happy & sad",
+    emoji: ":)",
+    keywords: [":-)", "happy", "smile", "joy"],
+  },
+  {
+    label: "Big Grin",
+    type: "Sy happy & sad",
+    emoji: ":D",
+    keywords: [":-D", "=D", "laugh", "big smile"],
+  },
+  {
+    label: "Wink",
+    type: "Sy happy & sad",
+    emoji: ";)",
+    keywords: [";-)", "wink", "flirt", "sarcasm"],
+  },
+  {
+    label: "Tongue",
+    type: "Sy happy & sad",
+    emoji: ":P",
+    keywords: [":-P", ":p", ":-p", "tongue", "silly", "playful"],
+  },
+  {
+    label: "Kiss",
+    type: "Sy happy & sad",
+    emoji: ":*",
+    keywords: [":-*", ":x", "kiss", "smooch", "love"],
+  },
+  {
+    label: "Heart",
+    type: "Sy happy & sad",
+    emoji: "<3",
+    keywords: ["heart", "love", "less than three"],
+  },
+
+  // --- Neutral/Skeptical ---
+  {
+    label: "unsure",
+    type: "Sy happy & sad",
+    emoji: ":/",
+    keywords: [":-/", ":\\", ":-\\", "skeptical", "unsure", "annoyed", "uneasy"],
+  },
+  {
+    label: "Neutral",
+    type: "Sy happy & sad",
+    emoji: ":|",
+    keywords: [":-|", "straight face", "indifferent", "no expression"],
+  },
+  {
+    label: "Shock",
+    type: "Sy happy & sad",
+    emoji: ":O",
+    keywords: [":-O", ":o", "shock", "surprise", "yell"],
+  },
+
+  // --- Negative/Sad ---
+  {
+    label: "Frown",
+    type: "Sy happy & sad",
+    emoji: ":(",
+    keywords: [":-(", "=(", "sad", "frown", "unhappy"],
+  },
+  {
+    label: "Crying",
+    type: "Sy happy & sad",
+    emoji: ":'(",
+    keywords: [":'-(", "cry", "tear", "sobbing"],
+  },
+  {
+    label: "Angry",
+    type: "Sy happy & sad",
+    emoji: ">:(",
+    keywords: [">:-(", "mad", "angry", "furious"],
+  },
+];
