+++ src/components/Records.tsx
@@ -12,7 +12,7 @@ const Records = ({ i, HandleCopy, PinHistory, removeHistory }: props) => {
   return (
     <div key={i.id} className="flex justify-between m-2 items-start">
       <div
-        className="p-3 mt-2 bg-blue-500/20  hover:bg-blue-500 hover:text-white w-75 max-h-27 line-clamp-4 overflow-hidden rounded-md cursor-pointer"
+        className="p-2 mt-2 bg-linear-to-r from-blue-500/20 from via-blue-700 to-blue-500/20  hover:bg-blue-500 hover:via-blue-500  hover:text-white h-20 w-75 line-clamp-4 overflow-hidden rounded-md cursor-pointer"
         onClick={() => HandleCopy(i.item)}
       >
         {i.item}
